requirements.txt
absl-py==2.4.0
appnope==0.1.4
asciidoc==10.2.1
asgiref==3.11.1
asttokens==3.0.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gi-docgen==2026.1
gunicorn==25.1.0
h5py==3.16.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.13.2
kiwisolver==1.5.0
libxml2-python==2.13.9
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
meson==1.10.1
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
opencv-python==4.13.0.92
optree==0.19.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pyzmq==27.1.0
regex==2026.2.28
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
smartypants==2.0.2
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
typogrify==2.1.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.12.0
